+++ apps/backend/src/user/entities/user.entity.ts
@@ -7,8 +7,8 @@ import {
     PrimaryGeneratedColumn,
     UpdateDateColumn,
 } from "typeorm";
-import { Job } from "./job.entity";
-import { Email } from "./email.entity";
+import { Job } from "../../job/entities/job.entity";
+import { Email } from "../../email/entities/email.entity";
 
 export enum UserRole {
     ADMIN = "admin",
